Fixed some tests after some small changes to collision enginen

+++ src/MarioAI/debugGraphics/DebugDraw.java
@@ -237,7 +237,7 @@ public class DebugDraw {
 			final ArrayList<Point> positions = new ArrayList<Point>(); 
 			for (DirectedEdge edge : path) {
 				for (Point2D.Float pos : edge.getMoveInfo().getPositions()) {
-					Point2D.Float correctPos = new Point2D.Float(edge.source.x + pos.x, edge.source.y - pos.y);
+					Point2D.Float correctPos = new Point2D.Float((float)edge.source.x + pos.x, edge.source.y - pos.y);
 					convertLevelPointToOnScreenPoint(observation, correctPos);
 					
 					positions.add(new Point((int)correctPos.x, (int)correctPos.y));
